Changed sticky to create a new stylesheet, rather than inserting rules into an existing one
On firefox, using CSSStyleSheet#insertRule raises a security error if
the stylesheets are not hosted on the same domain as the page. This is a
problem if you use a cdn to host your assets. Instead, I've made the
script insert a new stylesheet into the header with the sticky rules.
This works for me on all major browsers