Changed the height of the wedge to be outer height
Previously the wedge (put in place to stop the page shifting up when the stuck element becomes fixed) was the inner height of the stuck element. It should probably be the outer height instead (which includes margins). Otherwise there's a judder as the page height changes when the element becomes stuck/unstuck.